# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OverUnder API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/emissions/distribute": {
      "post": {
        "tags": ["emissions"],
        "summary": "Distribute",
        "description": "Distribute OU emissions from treasury to recipients.\n\nOperator-only. Calls EmissionsDistributor.distribute which transfers from treasury.\nNever mints. Program IDs: 0=LP, 1=maker, 2=agent, 3=quest.\nSigns with RELAYER_PRIVATE_KEY through the relayer's shared NonceManager (above the in-flight nonce\nfloor) so it never races the CLOB worker for a nonce. The contract has no idempotency key, so every\nsigned tx is recorded write-ahead: a retry of the same payload (and idempotencyKey) returns the\nrecorded distribution instead of paying twice, and an ambiguous broadcast answers 202 with its hash.",
        "operationId": "distribute_api_v1_emissions_distribute_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DistributeRequest" } } }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/emissions/distributions/{distribution_id}": {
      "get": {
        "tags": ["emissions"],
        "summary": "Get Distribution",
        "description": "Poll a distribution: moves sending/sent on from its receipt.",
        "operationId": "get_distribution_api_v1_emissions_distributions__distribution_id__get",
        "parameters": [
          {
            "name": "distribution_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Distribution Id" }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DistributeRequest": {
        "properties": {
          "program": { "type": "integer", "title": "Program" },
          "recipients": {
            "items": { "$ref": "#/components/schemas/Recipient" },
            "type": "array",
            "title": "Recipients"
          },
          "idempotencyKey": { "type": "string", "title": "Idempotencykey", "default": "" }
        },
        "type": "object",
        "required": ["program", "recipients"],
        "title": "DistributeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Recipient": {
        "properties": {
          "address": { "type": "string", "title": "Address" },
          "amount": { "type": "string", "title": "Amount" }
        },
        "type": "object",
        "required": ["address", "amount"],
        "title": "Recipient"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}